Remove old testing code
